Fix js build tag issue
