+++ internal/blob/blob_other.go
@@ -1,0 +1,5 @@
+// +build !js
+
+package blob
+
+type jsExtensions interface{}
